requirements.txt
asgiref==3.8.1
Django==4.1.13
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
django-environ==0.11.2
psycopg2==2.9.9
django-colorfield==0.11.0
pytest
pytest-django
pytest-sugar
djangorestframework
djangorestframework-simplejwt
